Address comments from Jakub
