--- operator-common/src/test/java/io/strimzi/operator/common/operator/resource/cruisecontrol/AbstractMockCruiseControl.java
@@ -1,62 +1,0 @@
-/*
- * Copyright Strimzi authors.
- * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
- */
-package io.strimzi.operator.common.operator.resource.cruisecontrol;
-
-import org.mockserver.configuration.ConfigurationProperties;
-import org.mockserver.integration.ClientAndServer;
-
-import java.io.ByteArrayInputStream;
-import java.io.File;
-import java.util.logging.LogManager;
-
-import static java.nio.charset.StandardCharsets.UTF_8;
-
-/**
- * Cruise Control mock.
- */
-public abstract class AbstractMockCruiseControl {
-    protected ClientAndServer server;
-
-    /**
-     * Sets up and returns a Cruise Control mock server.
-     *
-     * @param serverPort   The port number the server should listen on.
-     * @param tlsKeyFile   File containing the CA key.
-     * @param tlsCrtFile   File containing the CA crt.
-     *
-     * @return             The mock CruiseControl instance.
-     */
-    public AbstractMockCruiseControl(int serverPort, File tlsKeyFile, File tlsCrtFile) {
-        try {
-            ConfigurationProperties.logLevel("WARN");
-            ConfigurationProperties.certificateAuthorityPrivateKey(tlsKeyFile.getAbsolutePath());
-            ConfigurationProperties.certificateAuthorityCertificate(tlsCrtFile.getAbsolutePath());
-
-            String loggingConfiguration = "handlers=org.mockserver.logging.StandardOutConsoleHandler\n" +
-                "org.mockserver.logging.StandardOutConsoleHandler.level=WARNING\n" +
-                "org.mockserver.logging.StandardOutConsoleHandler.formatter=java.util.logging.SimpleFormatter\n" +
-                "java.util.logging.SimpleFormatter.format=%1$tF %1$tT  %3$s  %4$s  %5$s %6$s%n\n" +
-                ".level=" + ConfigurationProperties.javaLoggerLogLevel() + "\n" +
-                "io.netty.handler.ssl.SslHandler.level=WARNING";
-            LogManager.getLogManager().readConfiguration(new ByteArrayInputStream(loggingConfiguration.getBytes(UTF_8)));
-
-            this.server = new ClientAndServer(serverPort);
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    public void reset() {
-        server.reset();
-    }
-
-    public void stop() {
-        server.stop();
-    }
-    
-    public boolean isRunning() {
-        return server.isRunning();
-    }
-}
