--- operator-common/src/test/resources/example.json
@@ -1,1 +1,0 @@
-{"key": "value"}
