+++ test/src/main/java/io/strimzi/test/TestUtils.java
@@ -32,6 +32,7 @@ import java.io.PrintWriter;
 import java.io.Reader;
 import java.io.StringWriter;
 import java.net.ServerSocket;
+import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
@@ -45,6 +46,7 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.StringTokenizer;
@@ -56,6 +58,8 @@ import java.util.concurrent.TimeoutException;
 import java.util.function.BooleanSupplier;
 import java.util.function.Function;
 
+import static java.lang.String.format;
+import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Arrays.asList;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -561,5 +565,25 @@ public final class TestUtils {
         }
         file.deleteOnExit();
         return file;
+    }
+
+    /**
+     * Get JSON content as string from resource.
+     *
+     * @param resourcePath Resource path.
+     *
+     * @return JSON content as string.
+     */
+    public static String jsonFromResource(String resourcePath) {
+        try {
+            URI resourceURI = Objects.requireNonNull(TestUtils.class.getClassLoader().getResource(resourcePath)).toURI();
+            Optional<String> content = Files.lines(Paths.get(resourceURI), UTF_8).reduce((x, y) -> x + y);
+            if (content.isEmpty()) {
+                throw new IOException(format("File %s from resources was empty", resourcePath));
+            }
+            return content.get();
+        } catch (Throwable t) {
+            throw new RuntimeException(t);
+        }
     }
 }
