+++ topic-operator/src/main/java/io/strimzi/operator/topic/ReplicasChangeHandler.java
@@ -13,7 +13,6 @@ import io.strimzi.api.kafka.model.topic.KafkaTopicStatusBuilder;
 import io.strimzi.api.kafka.model.topic.ReplicasChangeStatusBuilder;
 import io.strimzi.operator.common.Reconciliation;
 import io.strimzi.operator.common.ReconciliationLogger;
-import io.strimzi.operator.common.Util;
 import io.strimzi.operator.common.model.cruisecontrol.CruiseControlEndpoints;
 import io.strimzi.operator.common.model.cruisecontrol.CruiseControlParameters;
 
@@ -267,7 +266,7 @@ public class ReplicasChangeHandler {
         if (config.cruiseControlAuthEnabled()) {
             String apiUsername = new String(getFileContent(config.cruiseControlApiUserPath()), UTF_8);
             String apiPassword = new String(getFileContent(config.cruiseControlApiPassPath()), UTF_8);
-            builder.header("Authorization", Util.buildBasicAuthValue(apiUsername, apiPassword));
+            builder.header("Authorization", TopicOperatorUtil.buildBasicAuthValue(apiUsername, apiPassword));
         }
     }
