+++ topic-operator/src/test/java/io/strimzi/operator/topic/TopicOperatorTestUtil.java
@@ -19,13 +19,21 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.junit.jupiter.api.TestInfo;
 
+import java.io.File;
+import java.io.IOException;
+import java.net.URI;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Predicate;
 import java.util.function.UnaryOperator;
 
 import static io.strimzi.test.k8s.KubeClusterResource.cmdKubeClient;
+import static java.lang.String.format;
+import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.junit.jupiter.api.Assertions.fail;
 
 class TopicOperatorTestUtil {
@@ -129,5 +137,18 @@ class TopicOperatorTestUtil {
 
     static KafkaTopic findKafkaTopicByName(List<KafkaTopic> topics, String name) {
         return topics.stream().filter(kt -> kt.getMetadata().getName().equals(name)).findFirst().orElse(null);
+    }
+    
+    static String stringFromFile(File filePath) {
+        try {
+            URI resourceURI = Objects.requireNonNull(filePath).toURI();
+            Optional<String> content = Files.lines(Paths.get(resourceURI), UTF_8).reduce((x, y) -> x + y);
+            if (content.isEmpty()) {
+                throw new IOException(format("File %s was empty", filePath.getAbsolutePath()));
+            }
+            return content.get();
+        } catch (Throwable t) {
+            throw new RuntimeException(t);
+        }
     }
 }
